+++ compiler/noirc_frontend/src/parser/parser/doc_comments.rs
@@ -1,4 +1,7 @@
-use crate::token::{DocStyle, Token, TokenKind};
+use crate::{
+    parser::ParserErrorReason,
+    token::{DocStyle, Token, TokenKind},
+};
 
 use super::{parse_many::without_separator, Parser};
 
@@ -27,6 +30,18 @@ impl<'a> Parser<'a> {
             | Token::BlockComment(comment, Some(DocStyle::Outer)) => comment,
             _ => unreachable!(),
         })
+    }
+
+    /// Skips any outer doc comments but produces a warning saying that they don't document anything.
+    pub(super) fn warn_on_outer_doc_comments(&mut self) {
+        let span_before_doc_comments = self.current_token_span;
+        let doc_comments = self.parse_outer_doc_comments();
+        if !doc_comments.is_empty() {
+            self.push_error(
+                ParserErrorReason::DocCommentDoesNotDocumentAnything,
+                span_before_doc_comments,
+            );
+        }
     }
 }
